+++ lithops/serverless/backends/code_engine/config.py
@@ -36,8 +36,10 @@ VALID_MEMORY_VALUES = [256, 512, 1024, 2048, 4096, 8192, 12288, 16384, 24576, 32
 VALID_REGIONS = ['us-south', 'us-east', 'ca-tor', 'eu-de', 'eu-gb', 'jp-osa', 'jp-tok', 'br-sao', 'au-syd']
 
 CLUSTER_URL = 'https://proxy.{}.codeengine.cloud.ibm.com'
+BASE_URL_V1 = 'https://api.{}.codeengine.cloud.ibm.com/api/v1'
+BASE_URL_V2 = 'https://api.{}.codeengine.cloud.ibm.com/v2'
 
-REQ_PARAMS = ('namespace',)
+REQ_PARAMS = ('iam_api_key',)
 
 DOCKERFILE_DEFAULT = """
 RUN apt-get update && apt-get install -y \
@@ -52,6 +54,7 @@ RUN pip install --upgrade --ignore-installed setuptools six pip \
         gevent \
         ibm-cos-sdk \
         ibm-vpc \
+        ibm-code-engine-sdk \
         redis \
         requests \
         PyYAML \
@@ -149,17 +152,17 @@ def load_config(config_data):
     if 'ibm' not in config_data or config_data['ibm'] is None:
         raise Exception("'ibm' section is mandatory in the configuration")
 
-    if 'iam_api_key' not in config_data['ibm']:
-        raise Exception("'iam_api_key' parameter is mandatory under the 'ibm' section of the configuration")
+    for param in REQ_PARAMS:
+        if param not in config_data['ibm']:
+            msg = f'"{param}" is mandatory in the "ibm" section of the configuration'
+            raise Exception(msg)
+
+    if not config_data['code_engine']:
+        config_data['code_engine'] = {}
 
     temp = copy.deepcopy(config_data['code_engine'])
     config_data['code_engine'].update(config_data['ibm'])
     config_data['code_engine'].update(temp)
-
-    for param in REQ_PARAMS:
-        if param not in config_data['code_engine']:
-            msg = f'"{param}" is mandatory in the "code_engine" section of the configuration'
-            raise Exception(msg)
 
     if 'region' not in config_data['code_engine']:
         msg = "'region' parameter is mandatory under the 'ibm' or 'code_engine' section of the configuration"