+++ lithops/standalone/backends/azure_vms/azure_vms.py
@@ -394,13 +394,13 @@ class AzureVMSBackend:
         """
         Builds a new VM Image
         """
-        pass
+        raise NotImplementedError()
 
     def list_images(self):
         """
         List VM Images
         """
-        pass
+        raise NotImplementedError()
 
     def _delete_vm_instances(self, all=False):
         """